requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.20

# AI/ML
openai==1.3.0
chromadb==0.4.18

# Testing (optional)
requests==2.31.0
